requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2